Adapt code and output to recalculate state totals
Previous counts combined hits for WK, GD, or PR. This was problematic since an issue that mentioned WK and GD would have been counted for each of those terms and then counted as 2 issues rather than a single one.

This new version looks to see if an issue uses any of the three terms and then counts that issue just once.